# backend/src/main/java/com/orthosmille/medic/modules/appointments/controller/AppointmentController.java
package com.orthosmille.medic.modules.appointments.controller;

import com.orthosmille.medic.modules.appointments.dto.AppointmentRequest;
import com.orthosmille.medic.modules.appointments.dto.AppointmentResponse;
import com.orthosmille.medic.modules.appointments.dto.AppointmentStatusRequest;
import com.orthosmille.medic.modules.appointments.service.AppointmentService;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/v1/appointments")
public class AppointmentController {

    private final AppointmentService appointmentService;

    public AppointmentController(AppointmentService appointmentService) {
        this.appointmentService = appointmentService;
    }

    @GetMapping
    public ResponseEntity<List<AppointmentResponse>> list() {
        return ResponseEntity.ok(appointmentService.findAll());
    }

    @GetMapping("/{id}")
    public ResponseEntity<AppointmentResponse> getById(@PathVariable Long id) {
        return ResponseEntity.ok(appointmentService.findById(id));
    }

    @PostMapping
    public ResponseEntity<AppointmentResponse> create(@Valid @RequestBody AppointmentRequest request) {
        return ResponseEntity.status(HttpStatus.CREATED).body(appointmentService.create(request));
    }

    @PutMapping("/{id}")
    public ResponseEntity<AppointmentResponse> update(@PathVariable Long id, @Valid @RequestBody AppointmentRequest request) {
        return ResponseEntity.ok(appointmentService.update(id, request));
    }

    @PatchMapping("/{id}/status")
    public ResponseEntity<AppointmentResponse> changeStatus(@PathVariable Long id, @Valid @RequestBody AppointmentStatusRequest request) {
        return ResponseEntity.ok(appointmentService.changeStatus(id, request));
    }
}