# backend/src/main/java/com/orthosmille/medic/modules/patients/controller/PatientController.java
package com.orthosmille.medic.modules.patients.controller;

import com.orthosmille.medic.modules.patients.dto.PatientRequest;
import com.orthosmille.medic.modules.patients.dto.PatientResponse;
import com.orthosmille.medic.modules.patients.service.PatientService;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/v1/patients")
public class PatientController {

    private final PatientService patientService;

    public PatientController(PatientService patientService) {
        this.patientService = patientService;
    }

    @GetMapping
    public ResponseEntity<List<PatientResponse>> list(@RequestParam(value = "q", required = false) String query) {
        return ResponseEntity.ok(patientService.findAll(query));
    }

    @GetMapping("/{id}")
    public ResponseEntity<PatientResponse> getById(@PathVariable Long id) {
        return ResponseEntity.ok(patientService.findById(id));
    }

    @PostMapping
    public ResponseEntity<PatientResponse> create(@Valid @RequestBody PatientRequest request) {
        return ResponseEntity.status(HttpStatus.CREATED).body(patientService.create(request));
    }

    @PutMapping("/{id}")
    public ResponseEntity<PatientResponse> update(@PathVariable Long id, @Valid @RequestBody PatientRequest request) {
        return ResponseEntity.ok(patientService.update(id, request));
    }

    @PatchMapping("/{id}/status")
    public ResponseEntity<PatientResponse> changeStatus(
            @PathVariable Long id,
            @RequestParam("active") boolean active
    ) {
        return ResponseEntity.ok(patientService.changeStatus(id, active));
    }
}